# src/main/java/com/sendroids/openapi/doc/test.yaml
openapi: 3.0.1
info:
  title: API Demo
  description: "这是一个简单的api"
  termsOfService: http://baidu.com
  contact:
    email: test@mail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8080/api
tags:
  - name: user-controller
    description: 用户相关
    externalDocs:
      description: 更多
      url: http://localhost:8080/api/user
paths:
  /user/auth:
    post:
      tags:
        - user-controller
      summary: 用户认证
      operationId: issueToken
      parameters:
        - name: username
          in: query
          description: 用户名
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: 密码
          required: true
          schema:
            type: string
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth"
        400:
          description: 错误的参数
          content: {}
        404:
          description: 错误的请求路径
          content: {}
        500:
          description: 服务器错误
          content: {}
  /user/{userId}:
    get:
      tags:
        - user-controller
      summary: 根据用户ID获取用户
      description: 返回一个用户对象
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          description: 用户ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: 错误的参数
          content: {}
        404:
          description: 错误的请求路径
          content: {}
        500:
          description: 服务器错误
          content: {}
  /user/addUser:
    post:
      tags:
        - user-controller
      summary: 用户注册
      operationId: addUser
      requestBody:
        description: 新增的用户对象
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        400:
          description: 错误的参数
          content: {}
        404:
          description: 错误的请求路径
          content: {}
        500:
          description: 服务器错误
          content: {}
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
          maxLength: 20
        password:
          type: string
          maxLength: 16
          minLength: 8
        email:
          type: string
          pattern: ^[a-zA-Z0-9_.-]+@[a-zA-Z0-9-]+(\.[a-zA-Z0-9-]+)*\.[a-zA-Z0-9]{2,6}$
        phone:
          type: string
          maxLength: 11
          minLength: 11
    Auth:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        token:
          type: string
    ResponseApi:
      type: object
      properties:
        code:
          type: integer
          format: int64
        message:
          type: string
        content:
          type: object
          items:
            type: object